+++ source/benoit/benoit.rs
@@ -23,3 +23,4 @@
 
 pub mod application;
 pub mod configuration;
+pub mod video;