+++ source/benoit/benoit/application/render_row.rs
@@ -23,22 +23,60 @@
 
 use crate::benoit::application::Application;
 
+extern crate rug;
+
+use rug::Float;
+use std::ops::{AddAssign, DivAssign, MulAssign};
+
 impl Application {
-	pub fn render_row(data: &mut [u32], y: u32, canvas_width: u32, canvas_height: u32, position_x: f64, position_y: f64, zoom: f64, maximum_iteration_count: u32) {
+	pub fn render_row(data: &mut [u32], precision: u32, y: u32, canvas_width: u32, canvas_height: u32, center_real: Float, center_imaginary: Float, zoom: Float, maximum_iteration_count: u32) {
+		let two = Float::with_val(precision, 2.0);
+
 		for x in 0x0..canvas_width {
-			let canvas_width  = canvas_width as f64;
-			let canvas_height = canvas_height as f64;
+			let canvas_width  = Float::with_val(precision, canvas_width);
+			let canvas_height = Float::with_val(precision, canvas_height);
 
-			let ca = (x as f64 - canvas_width  / 2.0) / canvas_width  * 4.0 / zoom + position_x;
-			let cb = (y as f64 - canvas_height / 2.0) / canvas_height * 4.0 / zoom + position_y;
+			//let ca = Float::with_val(precision, (&x - &canvas_width  / 2.0) / &canvas_width  * 4.0 / &zoom + &center_real);
+			//let cb = Float::with_val(precision, (&y - &canvas_height / 2.0) / &canvas_height * 4.0 / &zoom + &center_imaginary);
 
-			let mut za: f64 = 0.0;
-			let mut zb: f64 = 0.0;
+			let x_float = Float::with_val(precision, x);
+			let y_float = Float::with_val(precision, y);
+
+			// Re(c) = (x-canvas_width/2)/canvas_width*4/zoom+Re(z)
+
+			let ca = {
+				let tmp0 = Float::with_val(precision, &canvas_width / 2.0);
+
+				let mut ca = Float::with_val(precision, &x_float - &tmp0);
+				ca.div_assign(&canvas_width);
+				ca.mul_assign(4.0);
+				ca.div_assign(&zoom);
+				ca.add_assign(&center_real);
+
+				ca
+			};
+
+			// Im(c) = (x-canvas_height/2)/canvas_height*4/zoom+Im(z)
+
+			let cb = {
+				let tmp0 = Float::with_val(precision, &canvas_height / 2.0);
+
+				let mut cb = Float::with_val(precision, &y_float - &tmp0);
+				cb.div_assign(&canvas_height);
+				cb.mul_assign(4.0);
+				cb.div_assign(&zoom);
+				cb.add_assign(&center_imaginary);
+
+				cb
+			};
+
+			let mut za = Float::with_val(precision, &ca);
+			let mut zb = Float::with_val(precision, &cb);
 
 			let mut iteration_count: u32 = 0x0;
 			while iteration_count < maximum_iteration_count {
-				let square_distance = za * za + zb * zb;
-				if square_distance > 2.0 * 2.0 { break }
+				let square_distance = Float::with_val(precision, &za * &za + &zb * &zb);
+				if square_distance > 4.0 { break }
 
 				{
 					// z = z^2 + c
@@ -47,15 +85,19 @@ impl Application {
 					// a = a^2 - b^2
 					// b = 2abi
 
-					let za_temporary = za;
-					za = za * za - zb * zb + ca;
-					zb = za_temporary * zb * 2.0 + cb;
+					let za_temporary = Float::with_val(precision, &za);
+
+					za = Float::with_val(precision, &za * &za - &zb * &zb);
+					za = Float::with_val(precision, &za + &ca);
+
+					zb = Float::with_val(precision, &za_temporary * &zb);
+					zb = Float::with_val(precision, &zb * &two + &cb);
 				}
 
 				iteration_count += 0x1;
 			}
 
-			data[x as usize] = iteration_count;
+			unsafe { *data.get_unchecked_mut(x as usize) = iteration_count }
 		}
 	}
 }